Rename log to logf

+++ kernel/includes/logger.h
@@ -2,6 +2,6 @@
 #define __LOGGER_H__
 
 void init_logger();
-void log(const char* format, ...);
+void logf(const char* format, ...);
 
 #endif
